clean up selective_client

+++ selective_client/main.cpp
@@ -13,7 +13,8 @@ int main() {
     int server_socket = utils::socket_wrapper();
     struct sockaddr_in server_addr;
     auto params = utils::read_parameters("client.in");
-    utils::init_server_addr(server_addr, std::stoi((*params)[0]), std::stoi((*params)[1]));
+    utils::init_server_addr(server_addr, static_cast<uint32_t>(std::stoi((*params)[0])),
+                            static_cast<uint16_t>(std::stoi((*params)[1])));
     utils::set_connection_time_out(server_socket, CON_TIME_OUT);
     SelectiveClient client{server_socket, server_addr};
     std::string filename = (*params)[2];
